# server/swagger/jpes-v1.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Allstate JPES Server",
    "description": "Provides WAS inventory data",
    "contact": {
      "name": "PointSource"
    }
  },
  "basePath": "/api/v1",
  "schemes": [
    "http", "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/ping": {

      "x-handler": "ping",
      "get": {
        "summary": "Pings the server",
        "description": "Pings the server.  Used to determine if the server is in good health",
        "operationId": "handlePing",
        "responses": {
          "200": {
            "description": "Server is ok"
          }
        }
      }
    },
    "/waspackage": {
      "x-handler": "wasPackage",
      "get": {
        "summary": "Gets all was packages in the database",
        "description": "Gets all was packages in the database",
        "operationId": "query",
        "parameters": [
          {
            "name": "install_id",
            "in": "query",
            "description": "Install id",
            "type": "string",
            "required": false
          },
          {
            "name": "host",
            "in": "query",
            "description": "Host",
            "type": "string",
            "required": false
          },
          {
            "name": "location",
            "in": "query",
            "description": "Location",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Result set"
          }
        }
      },
      "post": {
        "summary": "Creates a new WAS package record",
        "description": "Creates a new WAS package record",
        "operationId": "create",
        "parameters": [
          {
            "name": "waspackage",
            "in": "body",
            "description": "New WAS package",
            "required": true,
            "schema": {
              "$ref": "./models/waspackage/waspackage.json"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/waspackage/{id}": {
      "x-handler": "wasPackage",
      "get": {
        "summary": "Gets all was packages in the database",
        "description": "Gets all was packages in the database",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "record id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Record"
          },
          "404" : {
            "description" : "Not found"
          }
        }
      },
      "put": {
        "summary": "Gets all was packages in the database",
        "description": "Gets all was packages in the database",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "record id",
            "type": "string",
            "required": true
          },
          {
            "name": "waspackage",
            "in": "body",
            "description": "Existing WAS package",
            "required": true,
            "schema": {
              "$ref": "./models/waspackage/waspackage.json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record"
          }
        }
      }
    }
  }
}